implemented Lobachevsky Method

+++ Solving-algebraic-equations-by-Lobachevsky-method/equation.h
@@ -6,18 +6,18 @@
 
 class Equation {
 public:
-    Equation(std::vector<double> coefs = {2.5, 0.3, -2, 1})
+    Equation(std::vector<long double> coefs = {2.5, 0.3, -2, 1})
         : m_coefs(coefs)
     {}
 
-    std::vector<double> getCoefs() const {return m_coefs;}
+    std::vector<long double> getCoefs() const {return m_coefs;}
 
-    double operator() (const double x) {
+    long double operator() (const long double x) {
         return (2.5 * pow(x, 3) + 0.3 * pow(x, 2) - 2 * x + 1);
     }
 
 private:
-    std::vector<double> m_coefs;
+    std::vector<long double> m_coefs;
 };
 
 #endif // EQUATION_H
